Trait jb::tool::action::Link

pub trait Link {
    // Required methods
    fn is_linked(&self) -> bool
       where Self: Sized;
    fn link(&self) -> Result<()>
       where Self: Sized;
    fn unlink(&self) -> Result<()>
       where Self: Sized;
    fn unlink_with_alternative(&self) -> Result<()>
       where Self: Sized;
}

Required Methods

fn is_linked(&self) -> bool
where Self: Sized,

Returns whether the tool is linked.

fn link(&self) -> Result<()>
where Self: Sized,

Links the tool.

This creates a symlink to the tool’s binary and icon in the PATH and icons directory, respectively.

Errors

This function will return an error if the tool is already linked, or if the symlinks fail.

fn unlink(&self) -> Result<()>
where Self: Sized,

Unlinks the tool.

This removes the symlink to the tool’s binary and icon in the PATH and icons directory, respectively.

Errors

This function will return an error if the tool is not linked, or if the symlinks fail.

fn unlink_with_alternative(&self) -> Result<()>
where Self: Sized,

Unlinks the tool and links an alternative version.

This removes the symlink to the tool’s binary and icon in the PATH and icons directory, respectively, and try to link an alternative version if available.

Errors

This function will return an error if the tool is not linked or if the symlinks fail.