Trait jb::tool::action::Probe

pub trait Probe {
    // Required method
    fn sync(&mut self) -> Result<Download>;
}

Required Methods

fn sync(&mut self) -> Result<Download>

Finds a compatible release for the tool.

Errors

This function will return an error if the release cannot be found.

Panics

This function will panic if the tool is not valid.